square root

+++ app/controllers/application_controller.rb
@@ -7,7 +7,7 @@ class ApplicationController < ActionController::Base
   def calculate_square
     # params = {"elephant" => "42"}
 
-    @num = params.fetch("elephant").to_i
+    @num = params.fetch("elephant").to_f
     @square_of_num = @num * @num
 
     render({ :template => "calculation_templates/square_results.html.erb"})
@@ -19,7 +19,7 @@ class ApplicationController < ActionController::Base
 
   def calculate_sqrt
     
-    @num = params.fetch("dog").to_i
+    @num = params.fetch("dog").to_f
     @squared_num = (@num ** 0.5).to_i
 
     render({ :template => "calculation_templates/sqrt_results.html.erb"})
